+++ src/scaffold/span.rs
